chore(refactor): cleaned up the long range comms to make the index.tsx file not so large
