+++ app/cards/LongRangeComm/longRangeCommPages.ts
@@ -1,0 +1,1 @@
+export type Pages = "inbox" | "sent" | "compose" | "outbox";
