+++ app/components/ui/icons/name.d.ts
@@ -1,99 +1,99 @@
 // This file is generated by icon spritesheet generator
 
 export const iconNames = [
-  "chevron-left",
+  "zap",
+  "x",
+  "wrench",
+  "volume-2",
+  "user-plus",
+  "unfold-vertical",
+  "text-cursor-input",
+  "sword",
+  "star",
+  "square-dashed",
+  "square-arrow-out-up-right",
+  "speaker",
+  "sparkles",
+  "siren",
+  "shuffle",
+  "shield-half",
+  "settings",
+  "send",
+  "route",
+  "rotate-cw",
+  "rotate-ccw",
+  "rocket",
+  "repeat-2",
+  "reactor",
+  "power-node",
+  "plus",
+  "plug",
+  "plug-zap",
+  "play",
+  "picture-in-picture",
+  "pencil",
+  "pencil-line",
+  "pause",
+  "package-open",
+  "orbit",
+  "octagon-alert",
   "minus",
+  "map",
+  "log-out",
+  "loader",
+  "joystick",
+  "info",
+  "inbox",
+  "home",
+  "hand",
+  "grip-vertical",
+  "graduation-cap",
+  "git-pull-request-create",
+  "git-branch",
+  "flame",
+  "file-video",
+  "file-up",
+  "file-type",
+  "file-text",
+  "file-audio",
+  "eye",
+  "external-link",
   "drafting-compass",
   "diamond-minus",
-  "octagon-alert",
-  "sword",
-  "atom",
-  "circle-off",
-  "flame",
-  "square-dashed",
-  "arrow-left-right",
-  "battery-75",
-  "git-pull-request-create",
-  "circle-dot",
-  "rocket",
-  "pause",
-  "circle-fading-arrow-up",
-  "file-type",
-  "grip-vertical",
-  "shuffle",
-  "atomic-slashes",
-  "circle-arrow-up",
-  "route",
-  "arrow-left",
-  "pencil-line",
-  "arrow-up",
-  "zap",
-  "external-link",
-  "eye",
-  "pencil",
-  "circle-arrow-down",
-  "battery-pack-alt",
-  "wrench",
-  "star",
-  "graduation-cap",
-  "camera",
-  "speaker",
-  "send",
-  "play",
-  "siren",
-  "battery-output",
-  "shield-half",
-  "loader",
-  "info",
-  "square-arrow-out-up-right",
-  "rotate-ccw",
-  "plug-zap",
-  "check",
-  "plus",
-  "reactor",
+  "database-zap",
+  "crosshair",
   "clipboard-list",
   "circle-slash",
-  "database-zap",
-  "git-branch",
-  "battery-0",
-  "plug",
-  "user-plus",
-  "orbit",
-  "archive",
-  "settings",
-  "log-out",
-  "brush",
-  "arrow-right",
-  "chevron-right",
-  "battery-25",
-  "repeat-2",
-  "ban",
-  "joystick",
-  "x",
-  "file-text",
-  "file-video",
-  "chevron-up",
-  "sparkles",
-  "inbox",
-  "map",
-  "battery-100",
-  "power-node",
-  "file-audio",
-  "rotate-cw",
-  "chevron-down",
-  "picture-in-picture",
-  "file-up",
-  "home",
-  "battery-input",
+  "circle-off",
+  "circle-fading-arrow-up",
+  "circle-dot",
+  "circle-arrow-up",
+  "circle-arrow-down",
   "chevrons-up-down",
+  "chevron-up",
+  "chevron-right",
+  "chevron-left",
+  "chevron-down",
+  "check",
+  "camera",
+  "brush",
+  "battery-pack-alt",
+  "battery-output",
+  "battery-input",
+  "battery-75",
   "battery-50",
-  "unfold-vertical",
-  "package-open",
-  "crosshair",
-  "text-cursor-input",
+  "battery-25",
+  "battery-100",
+  "battery-0",
+  "ban",
+  "atomic-slashes",
+  "atom",
+  "arrow-up",
+  "arrow-right",
+  "arrow-left",
+  "arrow-left-right",
   "arrow-down",
-  "volume-2",
-  "hand",
+  "archive",
 ] as const
 
 export type IconName = typeof iconNames[number]
